car_classification/requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# Progress Bar
tqdm>=4.65.0

# Optional: For visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
